requirements.txt
pytest==7.3.1
selene>=2.0.0rc4
allure-pytest